Move `getSchedulerIntegration` into generic utility
Fetching the API keys for the scheduler should be part of the generic
scheduler, not the post scheduler.

Git doesn't make it easy to see, but all I did was move
`getSchedulerIntegration` out of the `post-scheduling/` folder and into
`scheduling/utils.js`.
